Cleanup and better API handling on the downloads page
