+++ ui/src/components/FullPageError.tsx
@@ -2,17 +2,16 @@ import React, { ReactElement } from 'react';
 import { EmptyState, EmptyStateBody } from '@patternfly/react-core';
 
 type Props = {
+  title?: string;
   message: string;
 };
 
-export default function FullPageError({ message }: Props): ReactElement {
+export default function FullPageError({
+  title = 'There was an unexpected error!',
+  message,
+}: Props): ReactElement {
   return (
-    <EmptyState
-      status="danger"
-      variant="lg"
-      titleText="There was an unexpected error!"
-      headingLevel="h4"
-    >
+    <EmptyState status="danger" variant="lg" titleText={title} headingLevel="h4">
       <EmptyStateBody>{message}</EmptyStateBody>
     </EmptyState>
   );
