+++ ui/src/containers/DownloadsPage/UserServiceAccountToken.tsx
@@ -9,12 +9,12 @@ import {
   Title,
 } from '@patternfly/react-core';
 import { AxiosPromise } from 'axios';
-import { AlertCircle } from 'react-feather';
-import { ClipLoader } from 'react-spinners';
 
 import { V1TokenResponse, UserServiceApi } from 'generated/client';
 import useApiQuery from 'client/useApiQuery';
 import configuration from 'client/configuration';
+import FullPageSpinner from 'components/FullPageSpinner';
+import FullPageError from 'components/FullPageError';
 
 const userService = new UserServiceApi(configuration);
 
@@ -22,26 +22,6 @@ const fetchToken = (): AxiosPromise<V1TokenResponse> => userService.token({});
 
 export default function UserServiceAccountToken(): ReactElement {
   const { loading, error, data } = useApiQuery(fetchToken);
-
-  if (loading) {
-    return (
-      <div className="inline-flex items-center">
-        <ClipLoader size={16} color="currentColor" />
-        <span className="ml-2">Loading service account token...</span>
-      </div>
-    );
-  }
-
-  if (error || !data?.Token) {
-    return (
-      <div className="inline-flex items-center">
-        <AlertCircle size={16} />
-        <span className="ml-2">Unexpected error occurred while loading service account token</span>
-      </div>
-    );
-  }
-
-  const exportCommand = `export INFRA_TOKEN="${data.Token}"`;
 
   return (
     <Flex direction={{ default: 'column' }} spaceItems={{ default: 'spaceItemsMd' }}>
@@ -64,17 +44,28 @@ export default function UserServiceAccountToken(): ReactElement {
 
       <Divider component="div" />
 
-      <Title headingLevel="h3">Authenticating with infractl</Title>
-      <p>Run the following in a terminal to authenticate infractl for use:</p>
-      <ClipboardCopy
-        className="pf-v6-u-w-50-on-xl"
-        isReadOnly
-        hoverTip="Copy"
-        clickTip="Copied"
-        variant={ClipboardCopyVariant.expansion}
-      >
-        {exportCommand}
-      </ClipboardCopy>
+      {loading ? (
+        <FullPageSpinner title="Loading service account token..." />
+      ) : error || !data?.Token ? (
+        <FullPageError
+          title="There was an error loading the service account token, or the token was not returned by the server"
+          message={error?.message ?? ''}
+        />
+      ) : (
+        <>
+          <Title headingLevel="h3">Authenticating with infractl</Title>
+          <p>Run the following in a terminal to authenticate infractl for use:</p>
+          <ClipboardCopy
+            className="pf-v6-u-w-50-on-xl"
+            isReadOnly
+            hoverTip="Copy"
+            clickTip="Command copied!"
+            variant={ClipboardCopyVariant.expansion}
+          >
+            {`export INFRA_TOKEN="${data.Token}"`}
+          </ClipboardCopy>
+        </>
+      )}
     </Flex>
   );
 }
